Done a little bit more on load/save adding a line: theactin.loaddata((i/InterRecordIterations)); into the main loop after saving doesn't seem to disrupt the program wildly.  (though I haven't had time to do a long run to see if the results are identical)  Nor have I explicitly blanked the data on each run.
Loading is very slow, so prob best not to load on each timepoint.

Next step, I think, is to re-write the beginning of main() to parse the command line properly:  e.g. have letter at the beginning, say, R for Run, L for Load, etc. then be able to load from a certain point.  That should make it a little easier to test by just letting it go from a midpoint and letting it run, and making sure we haven't let any variables get through without saving since they'll be nutty values after the load if we do that and we should be able to pick them up.