Added some matlab '.m' files to process the node data. [The node data changed marginally recently, so some *may* need minor updates]
Thing I've been working on right now is the tracknodes.m routine to plot trajectories of nodes.  Syntax tracknodes(startframe, endframe).  Currently produces a static 3D plot of the node trajectories, and rotates the view.   The range of nodes to plot is specified in the function.

For some reason on my machine it's not saving to an avi file at the mo, and the movie seems to bounce back and forth across the screen when you play it (I think this is causing the avi probs).  Seems likely this is a matlab bug.  Not found workaround yet.