Fixed some sign errors in segments binning.  Also, replaced 'SEGMENT/2' with 'CAPSULE_HALF_LINEAR' etc. to avoid confusion.
Class 'segments':  Bins created so that there are RADIAL_SEGMENTS for one axis of sphere, or for both caps of the capsule, and the linear portion of the capsule has as close as possible to the same lengths of segments.  Segments line with edges of caps.  Lines plotted from middle of segments.  Nodes that impact the nucleator are passed via addsurfaceimpact() from nucleator::collision(), and sums are cleared from the main() loop with clearsurfaceimpacts().  drawoutline() and drawsurfaceimpacts() draw the outline and force lines by appending imagemagick draw commands to a passed stringstream.  drawsurfaceimpacts() has a scale parameter allowing overlapping draw commands to use different colors in actin::savebmp().  the segments object also stores segments radiating out from the surface of the nucleator for the linkforces, repulive forces etc. though this part isn't complete yet.  Also need to check fully that orientation of force bars matches the nucleator orientation.