VTK now invoked by 'comet vtk 100:100' (for making vtk jpeg etc) or 'comet vtkview 100:100' for interactive (and comet post 100:100 for bitmap)  (and removed VIS_INTERACTIVE from VTK)
Changed addGaussianNodeVolume() and added addStructuredPointIsoRender() so we can call it multiple times and have several surfaces with different thresholds, colours and opacities.

Added VTK_SCALECOLOPACITY to so that the alpha is scaled with the value as well as the colour

Changed bitmap plotting to allow REFERENCEFRAME (i.e. values are relative to that frame)

Fixed segments rotation for the new frame of ref (again) Hopefully it's right now.